Drop cssbeautify as dependency
The only thing is was changing was forcing inline comments on to a new line, and that's
not necessarily what's needed. Consider some CSS which has inline comments indicating a line
specifically for certain versions of IE. Moving those to next lines would make the code less clear.

The unit test therefore has been updated to test how CSSComb handles inline comments next to a property
using a space separator, a tab separator and not next to a property.

Fixes #3